refactor(courses): Group courses DTOs
