--- src/courses/domain/dtos/add_student_to_course_dto.go
@@ -1,7 +1,0 @@
-package dtos
-
-type AddStudentToCourseDTO struct {
-	TeacherUUID string
-	StudentUUID string
-	CourseUUID  string
-}
