--- src/courses/domain/dtos/enrolled_courses_dto.go
@@ -1,8 +1,0 @@
-package dtos
-
-import "github.com/UPB-Code-Labs/main-api/src/courses/domain/entities"
-
-type EnrolledCoursesDto struct {
-	Courses       []entities.Course
-	HiddenCourses []entities.Course
-}
