--- src/courses/domain/dtos/enrolled_students_dto.go
@@ -1,9 +1,0 @@
-package dtos
-
-type EnrolledStudentDTO struct {
-	UUID            string
-	FullName        string
-	Email           string
-	InstitutionalId string
-	IsActive        bool
-}
