--- src/courses/domain/dtos/get_invitation_code_dto.go
@@ -1,6 +1,0 @@
-package dtos
-
-type GetInvitationCodeDTO struct {
-	CourseUUID  string
-	TeacherUUID string
-}
