--- src/courses/domain/dtos/join_course_using_invitation_code_dto.go
@@ -1,6 +1,0 @@
-package dtos
-
-type JoinCourseUsingInvitationCodeDTO struct {
-	StudentUUID    string
-	InvitationCode string
-}
