--- src/courses/domain/dtos/rename_course_dto.go
@@ -1,7 +1,0 @@
-package dtos
-
-type RenameCourseDTO struct {
-	TeacherUUID string
-	CourseUUID  string
-	NewName     string
-}
